--- API/src/app/DSL/dsl.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { DSLController } from './controllers/dsl.controller';
-import { DSL } from './entities/dsl.entity';
-import { DSLService } from './services/dsl.service';
-
-@Module({
-    imports: [TypeOrmModule.forFeature([DSL])],
-    providers: [DSLService],
-    controllers: [DSLController],
-})
-export class DSLModule {}
