--- API/src/app/DSL/index.ts
@@ -1,5 +1,0 @@
-export * from './controllers/dsl.controller';
-export * from './entities/dsl.entity';
-export * from './repositeries/dsl.repository';
-export * from './services/dsl.service';
-export * from './dsl.module';
