--- API/src/app/DataPoint/controllers/dataPpoint.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller()
-export class DataPointController {}
