+++ API/src/app/DataPoint/dataPoint.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DataPointController } from './controllers/dataPpoint.controller';
+import { DataPointController } from './controllers/dataPoint.controller';
 import { DataPoint } from './entities/dataPoint.entity';
 import { DataPointService } from './services/dataPoint.service';
