+++ API/src/app/DatasetEntry/dataSetEntries.module.ts
@@ -1,0 +1,10 @@
+import { Module } from '@nestjs/common';
+import { DataSetEntryController } from './dataSetEntries.controller';
+import { DataSetEntriesService } from './dataSetEntries.service';
+
+@Module({
+    imports: [],
+    controllers: [DataSetEntryController],
+    providers: [DataSetEntriesService],
+})
+export class DataSetEntriesModule {}
