+++ API/src/app/DatasetEntry/dataSetEntries.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@nestjs/common';
+import { DataSetEntries } from './dataSetEntries.models';
+
+@Injectable()
+export class DataSetEntriesService {
+    getData(entriesQuery: DataSetEntries) {
+        return entriesQuery;
+    }
+}
