--- API/src/app/DatasetEntry/datasetEntry.module.ts
@@ -1,10 +1,0 @@
-import { Module } from '@nestjs/common';
-import { DataSetEntryController } from './datasetEntry.controller';
-import { DataSetEntryService } from './datasetEntry.service';
-
-@Module({
-    imports: [],
-    controllers: [DataSetEntryController],
-    providers: [DataSetEntryService],
-})
-export class DataSetEntriesModule {}
