--- API/src/app/DatasetEntry/datasetEntry.service.ts
@@ -1,9 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { DataSetEntry } from './datasetEntry.models';
-
-@Injectable()
-export class DataSetEntryService {
-    getData(entriesQuery: DataSetEntry) {
-        return entriesQuery;
-    }
-}
